requirements.txt
pdfplumber
PyYAML
streamlit
pytest
pandas
pdf2image
pytesseract
Pillow
